Relocated test_set_clim_emits_single_callback to test_colors.py
